webmastered/requirements.txt
anyascii==0.3.1
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.9.3
branca==0.5.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.6
django-allauth==0.51.0
django-browser-reload==1.6.0
django-debug-toolbar==3.5.0
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-redis==5.2.0
django-taggit==2.1.0
django-tailwind==3.3.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
folium==0.12.1.post1
hiredis==2.0.0
html5lib==1.1
idna==3.3
Jinja2==3.1.2
l18n==2021.3
MarkupSafe==2.1.1
numpy==1.23.1
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycountry==22.3.5
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
sendgrid==6.9.7
sentry-sdk==1.7.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
stripe==3.5.0
tablib==3.2.1
telepath==0.2
urllib3==1.26.10
wagtail==3.0.1
webencodings==0.5.1
Willow==1.4.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
